Move page component
