+++ src/Frontend/ClientApp/src/components/pages/App.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import reactLogo from "./assets/react.svg";
+import reactLogo from "../../assets/react.svg";
 import "./App.css";
 
 const App = (): JSX.Element => {
